+++ webstepper/src/index.tsx
@@ -1,30 +1,11 @@
-import React, { StrictMode, useState, Suspense } from "react";
+import React, { StrictMode, useState } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.js";
 import "./css/styles.scss";
+import "./i18n";
 
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { useMediaQuery, CssBaseline } from "@mui/material";
-
-// i18n configuration
-import i18n from "i18next";
-import { initReactI18next } from "react-i18next";
-import Backend from "i18next-http-backend";
-import LanguageDetector from "i18next-browser-languagedetector";
-
-i18n.use(Backend)
-    .use(LanguageDetector)
-    .use(initReactI18next)
-    .init({
-        fallbackLng: "en",
-        debug: process.env.NODE_ENV !== "production",
-        interpolation: {
-            escapeValue: false,
-        },
-        backend: {
-            loadPath: "/locales/{{lng}}/{{ns}}.json",
-        },
-    });
 
 declare module "@mui/material/styles" {
     interface PaletteColor {
@@ -154,8 +135,6 @@ const root = createRoot(document.getElementById("root"));
 
 root.render(
     <StrictMode>
-        <Suspense fallback={<div>Loading...</div>}>
-            <Root />
-        </Suspense>
+        <Root />
     </StrictMode>
 );