+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added "Theme" rendering option, allowing users to change the theme of output
 - Updated the default canvas bottom padding from 100 to 25
 - Extended dict `value` to accept a list of two-element lists, where each inner list is in the form `[key id, value id]`
-- Added internationalization (i18n) support with English and French translations across demo, webstepper, and CLI.
+- Added internationalization (i18n) support for translations across demo, webstepper, and CLI.
 
 ### 🐛 Bug fixes